Adaption to handle nested headers (see detailed description)
For future reference (explanation also present in code and git commits):

In order to be able to have full control of the columns in the presence of multiple header rows and with a “table-layout: fixed” css property, the following trick is applied:
the first row of header cells is a dummy row that has the same amount of cells as the number of columns. The css fixed table layout will take this row as the reference row to calculate the correct widths. If this row is not present, we have no control over the nested header cells. The styling of this row can be adapted such that it is not visible and needs no content (but the data-column-index attribute (starting from 0) needs to be set manually for every cell in this row).

A second extension is the fact that “data-columnname” and “data-parentcolumnname” must be used on top header cells and subheader cells to refer to each other.

(see https://www.w3.org/TR/CSS2/tables.html#fixed-table-layout and https://datatables.net/forums/discussion/39205/fixedheader-column-width-not-calculating-properly-with-multi-column-headers)